+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:push_button/push_button.dart';
 
 void main() {
   runApp(const MyApp());
@@ -11,12 +12,12 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
-      title: 'Flutter Demo',
+      title: 'Flutter Push Button Example',
       theme: ThemeData(
         colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
         useMaterial3: true,
       ),
-      home: const MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const MyHomePage(title: 'Flutter Push Button Example'),
     );
   }
 }
@@ -39,6 +40,18 @@ class _MyHomePageState extends State<MyHomePage> {
     });
   }
 
+  void _decrementCounter() {
+    setState(() {
+      _counter--;
+    });
+  }
+
+  void _resetCounter() {
+    setState(() {
+      _counter = 0;
+    });
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -57,12 +70,37 @@ class _MyHomePageState extends State<MyHomePage> {
               '$_counter',
               style: Theme.of(context).textTheme.headlineMedium,
             ),
-            const SizedBox(height: 20),
-            const Row(
-              children: [
-                // TODO: Add a button to increment and decrement the counter
-              ],
-            )
+            const SizedBox(height: 30),
+            PushButton(
+              backgroundImage: 'assets/images/wood.jpg',
+              borderRadius: BorderRadius.circular(30),
+              onPressed: _incrementCounter,
+              child: const Text(
+                'Increment',
+                style: TextStyle(color: Colors.white),
+              ),
+            ),
+            const SizedBox(height: 10),
+            PushButton(
+              onPressed: _decrementCounter,
+              child: const Text(
+                'Decrement',
+                style: TextStyle(color: Colors.white),
+              ),
+            ),
+            const SizedBox(height: 10),
+            PushButton(
+              backgroundColor: Colors.red,
+              onPressed: _resetCounter,
+              borderRadius: const BorderRadius.only(
+                topLeft: Radius.circular(30),
+                bottomRight: Radius.circular(30),
+              ),
+              child: const Text(
+                'Reset',
+                style: TextStyle(color: Colors.white),
+              ),
+            ),
           ],
         ),
       ),