+++ lib/push_button.dart
@@ -1,7 +1,3 @@
 library push_button;
 
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/push_button.dart';