+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 0.0.1
+## 0.0.1 - 2024-03-25
 
-* TODO: Describe initial release.
+- First release of push button ui package
+- Customizeable button shape
+- Customizeable button color and background
+- Support full image background easily just with adding image path to `backgroundImage` parameter.